Test - send a value transaction to local Tangle

+++ src/core/models/message.c
@@ -208,7 +208,7 @@ size_t core_message_serialize(core_message_t* msg, byte_t buf[], size_t buf_len)
   offset += sizeof(uint8_t);
 
   // parents
-  for (uint index = 0; index < parents_len; index++) {
+  for (uint8_t index = 0; index < parents_len; index++) {
     byte_t* parent_id = core_message_get_parent_id(msg, index);
     memcpy(buf + offset, parent_id, IOTA_MESSAGE_ID_BYTES);
     offset += IOTA_MESSAGE_ID_BYTES;
